+++ src/ServiceProvider/DevToolsServiceProvider.php
@@ -41,6 +41,10 @@ use FastForward\DevTools\Console\Logger\Processor\CommandOutputProcessor;
 use FastForward\DevTools\Console\Logger\Processor\CompositeContextProcessor;
 use FastForward\DevTools\Console\Logger\Processor\ContextProcessorInterface;
 use FastForward\DevTools\Console\Output\GithubActionOutput;
+use FastForward\DevTools\Console\Output\OutputCapabilityDetector;
+use FastForward\DevTools\Console\Output\OutputCapabilityDetectorInterface;
+use FastForward\DevTools\Environment\Environment;
+use FastForward\DevTools\Environment\EnvironmentInterface;
 use FastForward\DevTools\Filesystem\FinderFactory;
 use FastForward\DevTools\Filesystem\FinderFactoryInterface;
 use FastForward\DevTools\Filesystem\Filesystem;
@@ -69,8 +73,10 @@ use FastForward\DevTools\License\Resolver;
 use FastForward\DevTools\License\ResolverInterface;
 use FastForward\DevTools\PhpUnit\Coverage\CoverageSummaryLoader;
 use FastForward\DevTools\PhpUnit\Coverage\CoverageSummaryLoaderInterface;
+use FastForward\DevTools\Process\ColorPreservingProcessEnvironmentConfigurator;
 use FastForward\DevTools\Process\ProcessBuilder;
 use FastForward\DevTools\Process\ProcessBuilderInterface;
+use FastForward\DevTools\Process\ProcessEnvironmentConfiguratorInterface;
 use FastForward\DevTools\Process\ProcessQueue;
 use FastForward\DevTools\Process\ProcessQueueInterface;
 use FastForward\DevTools\Path\DevToolsPathResolver;
@@ -109,7 +115,10 @@ final class DevToolsServiceProvider implements ServiceProviderInterface
     {
         return [
             // Process
+            EnvironmentInterface::class => get(Environment::class),
+            OutputCapabilityDetectorInterface::class => get(OutputCapabilityDetector::class),
             ProcessBuilderInterface::class => get(ProcessBuilder::class),
+            ProcessEnvironmentConfiguratorInterface::class => get(ColorPreservingProcessEnvironmentConfigurator::class),
             ProcessQueueInterface::class => get(ProcessQueue::class),
 
             // Filesystem
@@ -145,7 +154,8 @@ final class DevToolsServiceProvider implements ServiceProviderInterface
                 ->method('setVerbosity', ConsoleOutputInterface::VERBOSITY_VERBOSE)
                 ->method('setFormatter', get(LogLevelOutputFormatter::class)),
             GithubActionOutput::class => create(GithubActionOutput::class)->constructor(
-                get(ConsoleOutputInterface::class)
+                get(ConsoleOutputInterface::class),
+                get(EnvironmentInterface::class)
             ),
             ContextProcessorInterface::class => create(CompositeContextProcessor::class)->constructor([
                 get(CommandInputProcessor::class),